+++ torch/_inductor/kernel/flex/__init__.py
@@ -1,0 +1,7 @@
+# mypy: allow-untyped-defs
+"""Flex attention kernel modules"""
+
+from .flex_attention import flex_attention, flex_attention_backward
+
+
+__all__ = ["flex_attention", "flex_attention_backward"]
